% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assrproj-package.R
\docType{package}
\name{assrproj-package}
\alias{assrproj-package}
\alias{assrproj}
\title{assrproj: Tools for ASSR analysis}
\description{
Utilities and pipelines for ITPC ~ exponent modeling, data prep, and plots.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
