% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{extractorRData}
\alias{extractorRData}
\title{Function for extracting an object from a .RData file created by R's save() command
Inputs: RData file, object name (optional)
If object is NULL or empty, returns the whole environment}
\usage{
extractorRData(file, object = NULL)
}
\description{
Function for extracting an object from a .RData file created by R's save() command
Inputs: RData file, object name (optional)
If object is NULL or empty, returns the whole environment
}
